# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StreamFlix API - Netflix-style streaming & movie discovery platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watch-progress/{itemId}:
    get:
      operationId: getWatchProgress
      tags: [watch-progress]
      summary: Get watch progress for a movie
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Watch progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchProgress"
    post:
      operationId: saveWatchProgress
      tags: [watch-progress]
      summary: Save or update watch progress
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchProgressInput"
      responses:
        "200":
          description: Saved watch progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchProgress"

  # ─── SEARCH ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    WatchProgress:
      type: object
      required: [itemId, positionSeconds, durationSeconds, percentage, completed]
      properties:
        id:
          type: integer
        userId:
          type: integer
        itemId:
          type: integer
        positionSeconds:
          type: integer
        durationSeconds:
          type: integer
        percentage:
          type: number
        completed:
          type: boolean
        lastWatchedAt:
          type: string
          format: date-time
    WatchProgressInput:
      type: object
      required: [positionSeconds, durationSeconds]
      properties:
        positionSeconds:
          type: integer
        durationSeconds:
          type: integer
...